% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_SD_cutoffs}
\alias{add_SD_cutoffs}
\title{Adds SD cutoffs to the data}
\usage{
add_SD_cutoffs(df, SD_cutoffs, sett)
}
\arguments{
\item{df}{original data}

\item{SD_cutoffs}{SD cutoffs for set of variables}

\item{sett}{list of settings}
}
\value{

}
\description{
Adds SD cutoffs to the data
}
\examples{
library(tidyverse)
library(baserater)

data(baserate_UDS)

sett <- create_settings(quo(gender),quo(edu_cat))

v <- controls \%>\% select(MMSE_total:BNT) \%>\% colnames()
sdi <- c(1,1,5)
controls2 <- multiply_df(controls, sdi)

compute_SD_cutoffs_limits(controls2,v, sett, by_gr1 = T, by_gr2 = T) \%>\%
add_SD_cutoffs(controls2, .,sett) 
}
