% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_group.R
\name{classify_group}
\alias{classify_group}
\title{Applies baserate approach to classify group into possible/probable}
\usage{
classify_group(df, v, df_controls, sdi = c(2, 1.5, 1), sett)
}
\arguments{
\item{df}{dataframe with group to classify}

\item{v}{which variables should be used for classification}

\item{df_controls}{data frame with controls used for computation of baserate scores}

\item{sdi}{vector of SDs for which we will classify the group when computing SD cutoffs}

\item{grouping_var1}{first quoted variable name, for which we break down the data}

\item{grouping_var2}{second quoted variable name, for which we break down the data}
}
\value{

}
\description{
Applies baserate approach to classify group into possible/probable
}
\examples{
library(dplyr)
library(baserater)
v <- c("Extraversion","Empathy.Agreeableness","Conscientiousness","Instability.Neuroticism","Openness_to_Experience.Intellect","Honesty.Humility",
      "Disinhibition", "Detachment", "Psychoticism","Negative_Affect","Antagonism")
df <- tibble() # here should be proper dataset, for example neuropsychology::personality
mutate_at(df, v, ~(-.))
sett <- create_settings(quo(Sex),quo(Meditation))
df_controls <- df \%>\% filter(Mood_Disorder == "Absence")
df_patients <- df \%>\% filter(Mood_Disorder == "Presence")
sdi <- c(2,1.5,1)
pat_classified <- classify_group(df_patients,v,df_controls,sdi,sett = sett)
}
