% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_nlower}
\alias{compute_nlower}
\title{Computes number of low scores}
\usage{
compute_nlower(df, v, sett)
}
\arguments{
\item{df}{data frame}

\item{v}{set of variables}

\item{sett}{list of settings}
}
\value{

}
\description{
Computes number of low scores
}
