% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_probposs_cutoffs.R
\name{compute_probposs_baserate_cutoffs}
\alias{compute_probposs_baserate_cutoffs}
\title{Compute baserate cutoff for division into possible and probable categories
Currently expects 0-2 variables that breaks the data
This function should be run after computing number of low scores.
The column name with lower scores is expected to be named n_lower}
\usage{
compute_probposs_baserate_cutoffs(
  df,
  v1 = NULL,
  v2 = NULL,
  possible = 0.8,
  probable = 0.9,
  round_dg = 3
)
}
\arguments{
\item{df}{tibble with data}

\item{v1}{first variable, that divide the data. Can be null, when we are only interested in overall data}

\item{v2}{second variable, that divide the data. Can be null, when we are only interested in one variable}

\item{possible}{- quantile, which is threshold for classification into possible}

\item{probable}{- quantile, which is threshold for classification into probable}

\item{round_dg}{to how many decimal places should we round}
}
\value{

}
\description{
Compute baserate cutoff for division into possible and probable categories
Currently expects 0-2 variables that breaks the data
This function should be run after computing number of low scores.
The column name with lower scores is expected to be named n_lower
}
\examples{
library(tidyverse)
library(baserater)
data(baserate_UDS)
v <- controls \%>\% select(MMSE_total:BNT) \%>\% colnames()
sett <- create_settings(quo(gender),quo(edu_cat))
sdi <- c(2,1.5,1)
controls2 <- multiply_df(controls, sdi)
compute_SD_cutoffs_limits(controls2,v, sett, by_gr1 = T, by_gr2 = T) \%>\%
  add_SD_cutoffs(controls2, .,sett) \%>\%
  test_if_lower(v,sett) \%>\%
  compute_nlower(v,sett) \%>\%
  compute_probposs_baserate_cutoffs(sett$grouping_var1, sett$grouping_var2, possible = .8, probable = .9)
  
}
