% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_SD_cutoffs.R
\name{compute_SD_cutoffs_limits}
\alias{compute_SD_cutoffs_limits}
\title{Computes SD limits for given SD}
\usage{
compute_SD_cutoffs_limits(df, v, sett, by_gr1 = F, by_gr2 = F)
}
\arguments{
\item{df}{data frame with variables}

\item{v}{vector of variable names, for which we compute the SD cutoffs}

\item{sdi}{SD scaling factor}

\item{grouping_var1}{first variable to create subgroups}

\item{grouping_var2}{second variable to create subgroups}

\item{lower_lim_suffix}{suffix for the variables that contain lower limits}
}
\value{

}
\description{
Computes SD limits for given SD
}
