% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_lower_lim_vars}
\alias{create_lower_lim_vars}
\title{For given variables, creates variable names for lower limits}
\usage{
create_lower_lim_vars(v, sett)
}
\arguments{
\item{uniq_suffix}{}
}
\value{

}
\description{
For given variables, creates variable names for lower limits
}
