% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_settings.R
\name{create_settings}
\alias{create_settings}
\title{Creates shared settings}
\usage{
create_settings(gr1, gr2)
}
\arguments{
\item{gr1}{quoted name of first variable}

\item{gr2}{quoted name of first variable}
}
\value{

}
\description{
Creates shared settings
}
\examples{
sett <- create_settings(quo(Age_categ),quo(Sex))
}
