% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_possible_probable.R
\name{evaluate_possible_probable}
\alias{evaluate_possible_probable}
\title{Evaluate data into possible and probable.
This function is not meant to be run outside the package}
\usage{
evaluate_possible_probable(df, v, sd_cutoffs, sett, possprob_suffix)
}
\arguments{
\item{df}{dataframe that should be evaluated into possible/probable}

\item{v}{which variables are used for evaluation}

\item{sd_cutoffs}{dataframe with cutoffs for each variable}

\item{possprob_suffix}{suffix for possible/probable}

\item{out_col}{name of ouotput column}

\item{join_by}{named vector that is supplied into join to merge data from cutoffs and evaluated data frame}

\item{is_lower_suffix}{suffix for is_lower_column}
}
\value{

}
\description{
Evaluate data into possible and probable.
This function is not meant to be run outside the package
}
