% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lower_lim}
\alias{lower_lim}
\title{Computes lower limit for subscale}
\usage{
lower_lim(x, sd_lim = 1.5, ...)
}
\arguments{
\item{x}{vector of values}

\item{sd_lim}{sd multiplier. Default value is 1.5}

\item{...}{additional parameters supplied to mean and sd functions. This is good way how to supply na.rm=T parameter}
}
\value{

}
\description{
Lower limit is compute by subtracting specified multiplier of SD from the mean
}
