% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{multiply_df}
\alias{multiply_df}
\title{Multiples data frame for given SD. The column name will be named as sdi}
\usage{
multiply_df(df, sdi)
}
\arguments{
\item{df}{data frame to multiply}

\item{sdi}{vector of SD scaling factors}
}
\value{

}
\description{
Multiples data frame for given SD. The column name will be named as sdi
}
