% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_SD_cutoffs}
\alias{add_SD_cutoffs}
\title{Adds SD cutoffs to the data}
\usage{
add_SD_cutoffs(df, SD_cutoffs, sett)
}
\arguments{
\item{df}{original data}

\item{SD_cutoffs}{SD cutoffs for set of variables}

\item{sett}{list of settings}
}
\value{

}
\description{
Adds SD cutoffs to the data
}
