% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_nlower}
\alias{compute_nlower}
\title{Computes number of low scores}
\usage{
compute_nlower(df, v, sett)
}
\arguments{
\item{df}{data frame}

\item{v}{set of variables}

\item{sett}{list of settings}
}
\value{

}
\description{
Computes number of low scores
}
\examples{
#' library(tidyverse)
library(baserater)

data(baserate_UDS)

sett <- create_settings(quo(gender),quo(edu_cat))

v <- controls \%>\% select(MMSE_total:BNT) \%>\% colnames()
sdi <- c(1,1,5)
controls2 <- multiply_df(controls, sdi)

compute_SD_cutoffs_limits(controls2,v, sett, by_gr1 = T, by_gr2 = T) \%>\%
add_SD_cutoffs(controls2, .,sett) \%>\%
compute_nlower(v,sett) \%>\%
}
