% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{test_if_lower}
\alias{test_if_lower}
\title{Tests, whether each of the variables is lower than appropriate lower limit}
\usage{
test_if_lower(df, v, sett)
}
\arguments{
\item{df}{data frame that should contain all variables "v" and all variable "v+lower_lim_sufix"}

\item{v}{vector of variable names used for computation of low scores}

\item{sett}{list of settings}
}
\value{

}
\description{
This function is used for computation of number of low scores. First, limits should be computed based on control sample. Second, each score is compared with the limit to measure, whether subject scored below limit.
}
